Fix structure issues
