+++ packages/frontend/src/components/Claim/Components/ClaimAmount/ClaimAmount.jsx
@@ -1,0 +1,53 @@
+import BlueButton from '../../../UI/button/BlueButton/BlueButton'
+import AnimationY from '@/components/Animated/Block/AnimationY/AnimationY'
+import { useCallback, useState } from 'react';
+import useWaveEffect from '@/hooks/useWaveEffect';
+import { ClaimAmountStyle } from './style';
+
+function ClaimAmount({ text, dataRadioButtons, buttonName, handleClick, wallet }) {
+    const [selectedRadioValue, setSelectedRadioValue] = useState(dataRadioButtons[0].value);
+
+    const handleRadioChange = useCallback((event) => {
+        setSelectedRadioValue(event.target.value);
+    },[])
+
+    return (
+        <ClaimAmountStyle>
+            <span className={'Amount'}>
+                <AnimationY teg={'p'} delay={100}>{text}</AnimationY>
+                {dataRadioButtons?.length
+                    ? dataRadioButtons.map((_, i) => (
+                        <RadioInput i={i} data={_} key={i} selectedRadioValue={selectedRadioValue} handleRadioChange={handleRadioChange}/>
+                    ))
+                    : null}
+            </span>
+            <AnimationY delay={(dataRadioButtons.length + 1) * 100 + 100}>
+                <BlueButton 
+                name={buttonName} 
+                handleClick={handleClick} 
+                disabled={wallet && wallet.length === 34 && selectedRadioValue ? false : true}
+                />
+            </AnimationY>
+        </ClaimAmountStyle>
+    )
+}
+
+function RadioInput({ i, data, selectedRadioValue, handleRadioChange }) {
+    const rippleRef = useWaveEffect(+selectedRadioValue === +data.value)
+
+    return (
+        <AnimationY delay={(i + 1) * 100 + 100} tag={'label'} key={i} className={'CustomRadio'}>
+            <input
+                type={'radio'}
+                name={'myRadio'}
+                value={data.value}
+                className={'RadioInput'}
+                onChange={handleRadioChange}
+                checked={+selectedRadioValue === +data.value}
+            />
+            <span ref={rippleRef} className={'RadioButton'}><p>{data.name}</p></span>
+        </AnimationY>
+    )
+}
+
+export default ClaimAmount
