+++ packages/frontend/src/components/Claim/Components/ClaimHeader/SvgAnimation.jsx
@@ -1,0 +1,17 @@
+function SvgAnimation() {
+    return (
+        <svg width="112" height="112" viewBox="0 0 112 112" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <g clipPath="url(#clip0_10681_90)">
+                <path d="M55.882 32V36.7044L69.3232 35.3603V42.0809L55.882 40.7368L52.6057 40.4113L52.5217 40.4008L52.4377 40.4113L49.1614 40.7368L35.7201 42.0809V35.3603L49.1614 36.7044V32H55.882ZM42.4408 52.1619L45.8011 48.8016H49.1614V44.1181L52.5217 43.7821L55.882 44.1181V48.8016H59.2423L62.6026 52.1619H65.9629C75.2458 52.1619 82.7645 59.6806 82.7645 68.9634V72.3237H69.3232V68.9634C69.3232 67.1048 67.8216 65.6031 65.9629 65.6031H62.1721C60.0509 68.6484 56.5121 70.6436 52.5217 70.6436C48.5313 70.6436 44.9925 68.6484 42.8713 65.6031H28.9995V52.1619H42.4408Z" fill="#008DE4" />
+            </g>
+            <path className={'Drop'} d="M78.7796 67.114L75.8603 62L72.941 67.114C72.647 67.618 72.5 68.1851 72.5 68.7626C72.5 70.5898 73.9806 72.0809 75.8183 72.0809H75.9023C77.7295 72.0809 79.2206 70.6003 79.2206 68.7626C79.2206 68.1851 79.0631 67.618 78.7796 67.114Z" fill="#008DE4" />
+            <defs>
+                <clipPath id="clip0_10681_90">
+                    <rect width="53.765" height="53.765" fill="white" transform="translate(29 25)" />
+                </clipPath>
+            </defs>
+        </svg>
+    )
+}
+
+export default SvgAnimation
