+++ packages/frontend/src/components/ui/button/BlueButton/BlueButton.jsx
@@ -1,6 +1,6 @@
 import Image from "next/image"
-import { BlueButtonStyle } from "./style"
 import useRipple from "@/hooks/useRipple";
+import { BlueButtonStyle } from "./style";
 
 function BlueButton({ name, handleClick, ariaLabel, altIconLeft = '', iconLeft, style, ...props }) {
     const rippleRef = useRipple(handleClick);
